move metrics code to dedicated files
